Update for Django 1.10 compatibility and repackage
* Add tox config for Django 1.10
* Move imports out of package __init__ to deal with django.core.exceptions.AppRegistryNotReady
* Fix up URLConf to be new format (list, with no prefix string)
* Update .gitignore
* Bump version to 0.3.0

All tests passing
