+++ .gitignore
@@ -2,3 +2,5 @@
 .idea/
 *.egg-info
 .coverage
+/htmlcov
+/.tox
