+++ net_promoter_score/__init__.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """Django app for managing Net Promoter Score data.
 
 The NPS is a measure of customer loyalty that is measured by asking your
@@ -28,19 +27,3 @@ each score is timestamped and linked to a Django User object, so you can
 easily work out the time elapsed since the last time they were asked.
 
 """
-from net_promoter_score import settings
-
-
-def show_nps(request):
-    """Return True if the NPS survey should be shown to the request user.
-
-    This is a dynamic function that uses the request object to determine
-    whether to display the survey. It defaults to a basic function that
-    looks up the last time the user was shown the survey, so that each
-    user is only shown it every X days. This should be overridden in the
-    Django settings to provide more sophisticated analysis - users with
-    different profiles may be surveyed more/less frequently, and you may
-    wish to survery based on previous answers.
-
-    """
-    return settings.NPS_DISPLAY_FUNCTION(request)
