+++ net_promoter_score/middleware.py
@@ -7,7 +7,7 @@ on each request. This value is added to the user session (so a max
 of one lookup per session.
 
 """
-from net_promoter_score import show_nps
+from net_promoter_score.utils import show_nps
 
 
 class NPSMiddleware(object):
