+++ net_promoter_score/urls.py
@@ -1,7 +1,12 @@
 # -*- coding: utf-8 -*-
-from django.conf.urls import patterns, url
+from django.conf.urls import url
 
-urlpatterns = patterns(
-    'net_promoter_score.views',
-    url(r'^score/$', 'post_score', name="post_score"),
-)
+from net_promoter_score import views
+
+urlpatterns = [
+    url(
+        r'^score/$',
+        views.post_score,
+        name="post_score"
+    ),
+]
