+++ requirements/django110.txt
@@ -1,0 +1,2 @@
+-r base.txt
+Django==1.10
