+++ setup.py
@@ -8,7 +8,7 @@ os.chdir(os.path.normpath(os.path.join(os.path.abspath(__file__), os.pardir)))
 
 setup(
     name="django-nps",
-    version="0.2.4-dev",
+    version="0.3.0",
     packages=find_packages(),
     install_requires=['Django>=1.8'],
     include_package_data=True,
