+++ tox.ini
@@ -1,5 +1,6 @@
 [tox]
-envlist = django18_py27
+envlist = django18_py27, django19_py27, django110_py27
+# Python 3 support is not quite there yet. Contributions welcome
 
 [testenv]
 basepython = python2.7
@@ -17,3 +18,10 @@ deps = -rrequirements/django19.txt
 [testenv:django19_py35]
 basepython = python3.5
 deps = -rrequirements/django19.txt
+
+[testenv:django110_py27]
+deps = -rrequirements/django110.txt
+
+[testenv:django110_py35]
+basepython = python3.5
+deps = -rrequirements/django110.txt
