+++ urls.py
@@ -1,13 +1,12 @@
 # -*- coding: utf-8 -*-
-from django.conf.urls import patterns, url, include
+from django.conf.urls import url, include
 from django.contrib import admin  # , staticfiles
 
 admin.autodiscover()
 
-urlpatterns = patterns(
-    '',
+urlpatterns = [
     url(r'^admin/', include(admin.site.urls)),
     url(r'^nps/', include('net_promoter_score.urls',
         namespace="net_promoter_score")),
     # url(r'^static/(?P<path>.*)$', staticfiles.views.serve),
-)
+]
